Fixed calls. Added Implementation for /pokedex get request call. Will return pokemon from List.

+++ src/main/java/com/example/pokemonapp/repository/PokemonRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PokemonRepository extends JpaRepository<Pokemon, Integer> {
+public interface PokemonRepository extends JpaRepository<Pokemon, Long> {
 }
